% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_modules}
\alias{genes_modules}
\title{Data frame of genes and modules}
\format{
A data frame with gene IDs in column 1 and modules in column 2.
}
\usage{
data(genes_modules)
}
\description{
Data frame of genes and modules
}
\examples{
data(genes_modules)
}
\keyword{datasets}
