% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hubs}
\alias{hubs}
\title{Hub genes}
\format{
A data frame with hub gene IDs in column 1, and module and degree
in columns 2 and 3.
}
\usage{
data(hubs)
}
\description{
Hub genes
}
\examples{
data(hubs)
}
\keyword{datasets}
