% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_enrich}
\alias{mod_enrich}
\title{Module enrichment results}
\format{
A data frame with module enrichment results for MapMan bins, 
GO terms and Interpro domains, as well as associated adjusted P-values.
}
\usage{
data(mod_enrich)
}
\description{
Module enrichment results
}
\examples{
data(mod_enrich)
}
\keyword{datasets}
