% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plotdata}
\alias{plotdata}
\title{Network plotting data}
\format{
A data frame with the x and y coordinates of nodes and edges between
them for plotting.
}
\usage{
data(plotdata)
}
\description{
Pre-computed data frame of network plotting data.
}
\examples{
data(plodata)
}
\keyword{datasets}
