% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scaled_degree}
\alias{scaled_degree}
\title{Data frame of genes and their scaled degree in the network}
\format{
A data frame with gene IDs in column 1 and scaled degree in column 2.
}
\usage{
data(scaled_degree)
}
\description{
Data frame of genes and their scaled degree in the network
}
\examples{
data(scaled_degree)
}
\keyword{datasets}
